copy the postgres-entrypoint.sh to docker-entrypoint-initdb.d
We can put your database initialization into the container's entrypoint directory: /docker-entrypoint-initdb.d. This executes all *.sh and *.sql files in this directory and does not touch the original command.
All files in this directory are automatically executed in the alphabetical order on container creation. Therefore, we can create a volume to add your scripts / sql files to the entrypoint and let the container execute them.